requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.6.0
dnspython==1.16.0
google-api-core==1.25.0
google-auth==1.24.0
google-cloud-dialogflow==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.34.1
idna==2.10
multidict==5.1.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.2
#python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
rsa==4.7
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
